Update to material 3 bottom buttons for input.
This required splitting up the button bar for compatability with the flutter API. Upgrading the design fixes #517.
