+++ app/lib/features/settings/export_import_screen.dart
@@ -3,8 +3,9 @@ import 'dart:io';
 import 'package:blood_pressure_app/components/disabled.dart';
 import 'package:blood_pressure_app/data_util/interval_picker.dart';
 import 'package:blood_pressure_app/features/export_import/active_field_customization.dart';
-import 'package:blood_pressure_app/features/export_import/export_button_bar.dart';
+import 'package:blood_pressure_app/features/export_import/export_button.dart';
 import 'package:blood_pressure_app/features/export_import/export_warn_banner.dart';
+import 'package:blood_pressure_app/features/export_import/import_button.dart';
 import 'package:blood_pressure_app/features/settings/tiles/dropdown_list_tile.dart';
 import 'package:blood_pressure_app/features/settings/tiles/input_list_tile.dart';
 import 'package:blood_pressure_app/features/settings/tiles/number_input_list_tile.dart';
@@ -180,7 +181,10 @@ class ExportImportScreen extends StatelessWidget {
             ],
           ),
         ),),
-      bottomNavigationBar: const ExportButtonBar(),
+      persistentFooterButtons: [
+        const ExportButton(),
+        const ImportButton(),
+      ],
     );
   }
 }
