+++ app/test/features/input/forms/date_time_form_test.dart
@@ -1,0 +1,24 @@
+import 'package:blood_pressure_app/model/storage/storage.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import '../../../util.dart';
+
+void main() {
+  testWidgets('saves entered values', (WidgetTester tester) async {
+
+  });
+
+  testWidgets('shows errors on bad inputs', (WidgetTester tester) async {
+
+  });
+
+  testWidgets('loads initial values', (WidgetTester tester) async {
+
+  });
+
+  testWidgets('saves only filled inputs', (WidgetTester tester) async {
+    
+  });
+}
